+++ clouddns/src/main/java/denominator/clouddns/CloudDNSResourceRecordSetApi.java
@@ -1,22 +1,32 @@
 package denominator.clouddns;
 
-import static denominator.clouddns.RackspaceApis.emptyOn404;
-import static denominator.common.Preconditions.checkNotNull;
-import static denominator.common.Util.filter;
-import static denominator.common.Util.nextOrNull;
-import static denominator.model.ResourceRecordSets.nameEqualTo;
-
-import java.net.URI;
-import java.util.Iterator;
-
-import javax.inject.Inject;
-
 import denominator.ResourceRecordSetApi;
 import denominator.clouddns.RackspaceApis.CloudDNS;
+import denominator.clouddns.RackspaceApis.JobIdAndStatus;
 import denominator.clouddns.RackspaceApis.ListWithNext;
 import denominator.clouddns.RackspaceApis.Pager;
 import denominator.clouddns.RackspaceApis.Record;
+import denominator.common.Util;
 import denominator.model.ResourceRecordSet;
+import feign.RetryableException;
+import feign.Retryer;
+
+import javax.inject.Inject;
+import java.net.URI;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
+import static denominator.clouddns.CloudDNSFunctions.toRDataMap;
+import static denominator.clouddns.RackspaceApis.emptyOn404;
+import static denominator.common.Preconditions.checkArgument;
+import static denominator.common.Preconditions.checkNotNull;
+import static denominator.common.Util.filter;
+import static denominator.common.Util.nextOrNull;
+import static denominator.common.Util.join;
+import static denominator.model.ResourceRecordSets.nameEqualTo;
 
 class CloudDNSResourceRecordSetApi implements denominator.ResourceRecordSetApi {
 
@@ -56,14 +66,92 @@ class CloudDNSResourceRecordSetApi implements denominator.ResourceRecordSetApi {
         return nextOrNull(new GroupByRecordNameAndTypeIterator(lazyIterateRecords(recordPager)));
     }
 
+    private static final int DEFAULT_TTL = 300;
+
     @Override
     public void put(ResourceRecordSet<?> rrset) {
-        throw new UnsupportedOperationException();
+        checkNotNull(rrset, "rrset was null");
+        checkArgument(!rrset.records().isEmpty(), "rrset was empty %s", rrset);
+
+        List<Map<String, Object>> recordsLeftToCreate = new ArrayList<Map<String, Object>>(rrset.records());
+
+        for (Record record : api.recordsByNameAndType(domainId, rrset.name(), rrset.type())) {
+            Map<String, Object> rdata = toRDataMap(record);
+
+            if (recordsLeftToCreate.contains(rdata)) {
+                recordsLeftToCreate.remove(rdata);
+
+                if (rrset.ttl() != null) {
+                    if (Util.equal(rrset.ttl(), record.ttl)) {
+                        continue;
+                    }
+
+                    awaitComplete(api.updateRecord(domainId, record.id, rrset.ttl(), record.data()));
+                }
+            } else {
+                awaitComplete(api.deleteRecord(domainId, record.id));
+            }
+        }
+
+        int ttlToApply = rrset.ttl() != null ? rrset.ttl() : DEFAULT_TTL;
+
+        for (Map<String, Object> rdata : recordsLeftToCreate) {
+            Map<String, Object> mutableRData = new LinkedHashMap<String, Object>(rdata);
+            Integer priority = getPriority(mutableRData);
+            String data = join(' ', mutableRData.values().toArray());
+
+            if (priority == null) {
+                awaitComplete(api.createRecord(domainId, rrset.name(), rrset.type(), ttlToApply, data));
+            } else {
+                awaitComplete(api.createRecordWithPriority(
+                        domainId, rrset.name(), rrset.type(), ttlToApply, data, priority));
+            }
+        }
+    }
+
+    private void awaitComplete(JobIdAndStatus job) {
+        RetryableException retryableException = new RetryableException(
+                String.format("JobId %s did not complete. Check your logs.", job.id), null);
+        Retryer retryer = new Retryer.Default(500, 1000, 30);
+
+        while (true) {
+            JobIdAndStatus jobIdAndStatus = api.getStatus(job.id);
+
+            if ("COMPLETED".equals(jobIdAndStatus.status)) {
+                break;
+            } else if ("ERROR".equals(jobIdAndStatus.status)) {
+                throw retryableException;
+            }
+
+            retryer.continueOrPropagate(retryableException);
+        }
+    }
+
+    /**
+     * Has the side effect of removing the priority from the mutableRData.
+     *
+     * @return null or the priority, if it exists for a MX or SRV record
+     */
+    private Integer getPriority(Map<String, Object> mutableRData) {
+        Integer priority = null;
+
+        if (mutableRData.containsKey("priority")) { // SRVData
+            priority = Integer.class.cast(mutableRData.remove("priority"));
+        } else if (mutableRData.containsKey("preference")) { // MXData
+            priority = Integer.class.cast(mutableRData.remove("preference"));
+        }
+
+        return priority;
     }
 
     @Override
     public void deleteByNameAndType(String name, String type) {
-        throw new UnsupportedOperationException();
+        checkNotNull(name, "name");
+        checkNotNull(type, "type");
+
+        for (Record record : api.recordsByNameAndType(domainId, name, type)) {
+            awaitComplete(api.deleteRecord(domainId, record.id));
+        }
     }
 
     static final class Factory implements denominator.ResourceRecordSetApi.Factory {
@@ -83,8 +171,10 @@ class CloudDNSResourceRecordSetApi implements denominator.ResourceRecordSetApi {
 
     Iterator<Record> lazyIterateRecords(final Pager<Record> recordPager) {
         final ListWithNext<Record> first = emptyOn404(recordPager, null);
+
         if (first.next == null)
             return first.iterator();
+
         return new Iterator<Record>() {
             Iterator<Record> current = first.iterator();
             URI next = first.next;